Move Symphony-specific models into app/models/symphony
