+++ spec/controllers/mediated_pages_controller_spec.rb
@@ -78,8 +78,8 @@ RSpec.describe MediatedPagesController do
       end
 
       it 'is allowed if the library ID field is filled out' do
-        allow(Patron).to receive(:find_by).with(library_id: '12345').and_return(
-          instance_double('Patron', email: nil, exists?: true)
+        allow(Symphony::Patron).to receive(:find_by).with(library_id: '12345').and_return(
+          instance_double(Symphony::Patron, email: nil, exists?: true)
         )
 
         put :create, params: {
