+++ spec/controllers/pages_controller_spec.rb
@@ -89,8 +89,8 @@ describe PagesController do
       end
 
       it 'is allowed if the library ID field is filled out' do
-        allow(Patron).to receive(:find_by).with(library_id: '12345').and_return(
-          instance_double('Patron', email: nil, exists?: true)
+        allow(Symphony::Patron).to receive(:find_by).with(library_id: '12345').and_return(
+          instance_double(Symphony::Patron, email: nil, exists?: true)
         )
 
         put :create, params: {
@@ -198,8 +198,8 @@ describe PagesController do
       end
 
       it 'redirects to success page with token when the sso user supplies a library ID' do
-        allow(Patron).to receive(:find_by).with(library_id: '5432123').and_return(
-          instance_double('Patron', email: nil, exists?: true)
+        allow(Symphony::Patron).to receive(:find_by).with(library_id: '5432123').and_return(
+          instance_double(Symphony::Patron, email: nil, exists?: true)
         )
 
         post :create, params: {
