+++ spec/features/status_page_spec.rb
@@ -27,7 +27,8 @@ describe 'Status Page' do
     let(:user) { create(:library_id_user) }
 
     before do
-      allow(Patron).to receive(:find_by).with(library_id: user.library_id).and_return(instance_double(Patron, exists?: true))
+      allow(Symphony::Patron).to receive(:find_by).with(library_id: user.library_id).and_return(instance_double(Symphony::Patron,
+                                                                                                                exists?: true))
     end
 
     it 'is available' do
