+++ spec/jobs/submit_borrow_direct_request_job_spec.rb
@@ -9,7 +9,7 @@ describe SubmitBorrowDirectRequestJob, type: :job do
 
   before do
     Sidekiq.logger.level = Logger::UNKNOWN
-    allow(Patron).to receive(:find_by).with(library_id: user.library_id).at_least(:once).and_return(
+    allow(Symphony::Patron).to receive(:find_by).with(library_id: user.library_id).at_least(:once).and_return(
       double(exists?: true, email: nil)
     )
     allow(request).to receive(:searchworks_item).and_return(sw_item)
