+++ spec/jobs/submit_iplc_listener_job_spec.rb
@@ -9,7 +9,7 @@ describe SubmitIplcListenerJob, type: :job do
 
   before do
     Sidekiq.logger.level = Logger::UNKNOWN
-    allow(Patron).to receive(:find_by).with(library_id: user.library_id).at_least(:once).and_return(
+    allow(Symphony::Patron).to receive(:find_by).with(library_id: user.library_id).at_least(:once).and_return(
       double(exists?: true, email: 'patron@example.com')
     )
     allow(request).to receive(:searchworks_item).and_return(sw_item)
