+++ spec/models/requests/hold_recall_spec.rb
@@ -24,7 +24,7 @@ describe HoldRecall do
       let(:subject) { create(:hold_recall, user: user) }
 
       before do
-        allow(Patron).to receive(:find_by).with(library_id: user.library_id).at_least(:once).and_return(
+        allow(Symphony::Patron).to receive(:find_by).with(library_id: user.library_id).at_least(:once).and_return(
           double(exists?: true, email: nil)
         )
       end
